Use base class for Jaguar Ipace battery
