+++ Software/src/battery/BATTERIES.h
@@ -2,19 +2,6 @@
 #define BATTERIES_H
 #include "../../USER_SETTINGS.h"
 
-#include "src/devboard/utils/types.h"
-
-// Abstract base class for next-generation battery implementations.
-// Defines the interface to call battery specific functionality.
-// No support for double battery yet.
-class CanBattery {
- public:
-  virtual void setup(void) = 0;
-  virtual void handle_incoming_can_frame(CAN_frame rx_frame) = 0;
-  virtual void update_values() = 0;
-  virtual void transmit_can() = 0;
-};
-
 #ifdef BMW_SBOX
 #include "BMW-SBOX.h"
 void handle_incoming_can_frame_shunt(CAN_frame rx_frame);
