+++ Software/src/battery/RENAULT-ZOE-GEN1-BATTERY.h
@@ -1,11 +1,10 @@
 #ifndef RENAULT_ZOE_GEN1_BATTERY_H
 #define RENAULT_ZOE_GEN1_BATTERY_H
-#include "../include.h"
+
+#include "CanBattery.h"
 
 #define BATTERY_SELECTED
-
-// Indicates that the object-oriented battery interface is to be activated.
-#define OO_BATTERY_SELECTED
+#define SELECTED_BATTERY_CLASS RenaultZoeGen1Battery
 
 #define MAX_PACK_VOLTAGE_DV 4200  //5000 = 500.0V
 #define MIN_PACK_VOLTAGE_DV 3000
